requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
datasets==2.18.0
dill==0.3.8
distlib==0.4.0
filelock==3.20.3
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
identify==2.6.16
idna==3.11
kaggle==1.8.3
kagglehub==0.4.1
kagglesdk==0.1.15
librt==0.7.8
multidict==6.7.1
multiprocess==0.70.16
multitasking==0.0.12
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pathspec==1.0.3
peewee==3.19.0
platformdirs==4.5.1
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.4
pyarrow==23.0.0
pyarrow-hotfix==0.7
pycparser==3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
text-unidecode==1.3
tqdm==4.67.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
types-tqdm==4.67.0.20250809
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
webencodings==0.5.1
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
yfinance==1.1.0
scikit-learn==1.8.0
chronos-forecasting>=2.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
boto3>=1.34.0
torch>=2.0.0
scikit-learn
torch
transformers
streamlit
chronos-forecasting
